fix bignum for tlb

+++ tlb/src/constructor/IntTlbConstructor.kt
@@ -1,6 +1,7 @@
 package org.ton.tlb.constructor
 
 import org.ton.bigint.BigInt
+import org.ton.bigint.toLong
 import org.ton.cell.CellBuilder
 import org.ton.cell.CellSlice
 import org.ton.cell.invoke
@@ -26,13 +27,13 @@ public class IntTlbConstructor(
 
     public companion object {
         public fun byte(length: Int = Byte.SIZE_BITS): TlbConstructor<Byte> =
-            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toByte() })
+            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toLong().toByte() })
 
         public fun short(length: Int = Short.SIZE_BITS): TlbConstructor<Short> =
-            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toShort() })
+            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toLong().toShort() })
 
         public fun int(length: Int = Int.SIZE_BITS): TlbConstructor<Int> =
-            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toInt() })
+            number(encode = { storeInt(it, length) }, decode = { loadInt(length).toLong().toInt() })
 
         public fun long(length: Int = Long.SIZE_BITS): TlbConstructor<Long> =
             number(encode = { storeInt(it, length) }, decode = { loadInt(length).toLong() })
